Fixed End Screen Issues
Added ability to maintain score from scene to scene
