+++ Totally  Accurate  Gladiator  Arena Game Files/Assets/Prefabs/UIPrefabs/UI.prefab
@@ -9344,7 +9344,7 @@ GameObject:
   - component: {fileID: 8983849450093721110}
   - component: {fileID: 8983849450093721105}
   - component: {fileID: 8983849450093721104}
-  - component: {fileID: 1073071388}
+  - component: {fileID: 18209158601577920}
   m_Layer: 0
   m_Name: TestingV2
   m_TagString: Untagged
@@ -9383,7 +9383,6 @@ MonoBehaviour:
   UF: {fileID: 8983849450093721110}
   US: {fileID: 8983849450093721105}
   UST: {fileID: 8983849450093721104}
-  UE: {fileID: 1073071388}
   StatueStatus: 
 --- !u!114 &8983849450093721101
 MonoBehaviour:
@@ -9672,7 +9671,7 @@ MonoBehaviour:
   SA: {fileID: 8983849450093721102}
   CapturedPoints: 0
   KilledPoints: 0
---- !u!114 &1073071388
+--- !u!114 &18209158601577920
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -9681,28 +9680,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 8983849450093721088}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9a3836997b6c34045824d09c299e104d, type: 3}
+  m_Script: {fileID: 11500000, guid: 49aa5e8434d77ab4babe929b3ed5c5b7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  UI_Finish: {fileID: 452056181}
-  Section1: {fileID: 8983849448398445482}
-  Section2: {fileID: 8983849449854606689}
-  Section3: {fileID: 8983849448714137457}
-  Section4: {fileID: 8983849450037620138}
-  Section5: {fileID: 8983849450061706691}
-  Points: {fileID: 2141335732}
-  Winner: {fileID: 1497318183}
-  teams: 
-  Cavemen: 0
-  Romans: 0
-  Vikings: 0
-  Knights: 0
-  Gamers: 0
-  cavs: {fileID: 1629610316338642795, guid: 5876854d8903749478eb955ea61374eb, type: 3}
-  roms: {fileID: 2092719305420882267, guid: a69cec5cf1cb65046b9e1a7d455356e9, type: 3}
-  viks: {fileID: 3256804404424912927, guid: a1859f21ab527bf4d81a3a0ab4a98a6c, type: 3}
-  kigs: {fileID: 5099182354126668163, guid: 22e53f1e388f15444b0daa1e768ddb24, type: 3}
-  gams: {fileID: 2902347704254253354, guid: ba78b2c1f0d797e45865b249bf0406f8, type: 3}
 --- !u!1 &8983849450118735342
 GameObject:
   m_ObjectHideFlags: 0
