+++ Totally  Accurate  Gladiator  Arena Game Files/Assets/Scenes/EndScene.unity
@@ -188,63 +188,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 28da87eb12161474796695a3465a9c39, type: 3}
---- !u!1001 &1542992999
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_RootOrder
-      value: 9
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839284, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8983849448635839285, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
-      propertyPath: m_Name
-      value: UI
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 6d0f2d7cd7787c44dbdb797cf5b66b37, type: 3}
 --- !u!1 &1977024243
 GameObject:
   m_ObjectHideFlags: 0
@@ -842,6 +785,63 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 22e53f1e388f15444b0daa1e768ddb24, type: 3}
+--- !u!1001 &6397298283492987224
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2624138007114929674, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_Name
+      value: End UI
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_RootOrder
+      value: 9
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2624138007114929675, guid: 4465317827371424ba7023f1fbf38285, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 4465317827371424ba7023f1fbf38285, type: 3}
 --- !u!1001 &7658862734046732525
 PrefabInstance:
   m_ObjectHideFlags: 0
