+++ Totally  Accurate  Gladiator  Arena Game Files/Assets/Scripts/UI2.0/UITestV2.cs
@@ -153,7 +153,7 @@ public class UITestV2 : MonoBehaviour
         kknights = TD.OtherTeamsKilled["Knights"]["Cavemen"];
         kvikings = TD.OtherTeamsKilled["Knights"]["Vikings"];
 
-        kills = TD.ScoreInfo["Cavemen"][0];
-        captures = TD.ScoreInfo["Cavemen"][1];
+        //kills = TD.ScoreInfo["Cavemen"][0];
+        //captures = TD.ScoreInfo["Cavemen"][1];
     }
 }
