+++ Totally  Accurate  Gladiator  Arena Game Files/Assets/Scripts/UI2.0/UI_Timer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49aa5e8434d77ab4babe929b3ed5c5b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
